Split react player props and mouse activity
